Enabled 64-bit architectures in Xcode project; fixed unit test functionality broken in 64-bit. Updated unit tests to avoid type mismatch errors in test macros (e.g. unsigned/NSUInteger).
git-svn-id: http://dysart.cs.byu.edu/chsvn/trunk@290 a75b6edf-6070-4da8-89c8-9992cc802e51

+++ test/CHAVLTreeTest.m
@@ -97,7 +97,7 @@ static NSMutableString *balanceErrors;
 - (void) testAddObject {
 	STAssertThrows([tree addObject:nil], @"Should raise an exception.");
 	
-	STAssertEquals([tree count], 0u, @"Incorrect count.");
+	STAssertEquals([tree count], (NSUInteger)0, @"Incorrect count.");
 	for (id object in objects)
 		[tree addObject:object];
 	STAssertEquals([tree count], [objects count], @"Incorrect count.");
